requirements.txt
asgiref==3.5.2
astroid==2.12.11
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
black==22.10.0
click==8.1.3
colorama==0.4.5
coverage==6.5.0
decorator==5.1.1
dill==0.3.5.1
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
executing==1.1.1
inflection==0.5.1
ipdb==0.13.9
ipython==8.5.0
isort==5.10.1
jedi==0.18.1
jsonschema==4.16.0
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.1
pep8==1.7.1
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.4
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
traitlets==5.5.0
tzdata==2022.5
uritemplate==4.1.1
uuid==1.30
wcwidth==0.2.5
wrapt==1.14.1
yapf==0.32.0
